**Summary** 

* Compiled using Matlab2023b in Apple M1 chip. 
* Remove debuging message (large improve of performance) 

**Compiled using :** 

>> mex -largeArrayDims -compatibleArrayDims dg_voronoi.c

**Time to compute Voronoi:** 

Without debug message: 
Matlab 2023b (macOs native): Elapsed time is 0.451528 seconds.
Matlab 2021b  (macOs roseta): Elapsed time is 0.636993 seconds.
Matlab 2021b ( Linux):  Elapsed time is 0.289606 seconds.


With debug message: 
Matlab 2021b (macOs):  Elapsed time is 198.217323 seconds.

**Todo:** 

- [x] Recompile on Windows 
- [x] Recompile on Linux
- [x] Recompile on MacOs
- [x] Improve progress bar management